split SyncToken into its own module
