+++ apps/server/src/routers/kobo/mod.rs
@@ -3,6 +3,7 @@ use axum::Router;
 use crate::config::state::AppState;
 
 mod sync;
+mod sync_token;
 
 pub(crate) fn mount(app_state: AppState) -> Router<AppState> {
 	Router::new().nest("/kobo", sync::mount(app_state))
