+++ apps/server/src/routers/kobo/sync_token.rs
@@ -1,0 +1,88 @@
+// The sync token is an string that is used to track the client's progress through the
+// sync. It is generated by the server and opaque to the client, so we can use it to
+// store any necessary information we need.
+//
+// It is generated by the server and sent to the client in a /v1/library/sync response,
+// in the "x-kobo-synctoken" header.
+//
+// The client passes the value back to the server in its next /v1/library/sync request,
+// in the "x-kobo-synctoken" header.
+use axum::http::HeaderValue;
+use base64::{engine::general_purpose::STANDARD_NO_PAD as BASE64, Engine};
+use std::str::Utf8Error;
+
+use reqwest::header::{InvalidHeaderValue, ToStrError};
+use serde::{Deserialize, Serialize};
+use thiserror::Error;
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct SyncToken {
+	// allows us to change the structure of the token in the future.
+	version: u64,
+
+	// the ID of the database KoboSync
+	pub sync_id: String,
+
+	// has the client retrieved all the data that was available in this session?
+	// tracking this explicitly should make it easier to support paging through proxied Kobo Store
+	// responses.
+	pub completed: bool,
+
+	// the offset of the next page that should be sent to the client.
+	// this is not meaningful if complete is true.
+	pub next_offset: usize,
+}
+
+#[derive(Error, Debug)]
+pub enum SyncTokenSerializeError {
+	#[error("Could not serialize JSON: {0}")]
+	JSONError(#[from] serde_json::Error),
+	#[error("Could not encode this token as a header: {0}")]
+	InvalidHeaderError(#[from] InvalidHeaderValue),
+}
+
+#[derive(Error, Debug)]
+pub enum SyncTokenDeserializeError {
+	#[error("Could not deserialize string from header: {0}")]
+	HeaderToStrError(#[from] ToStrError),
+	#[error("Could not decode UTF-8: {0}")]
+	UTF8Error(#[from] Utf8Error),
+	#[error("Could not decode Base64: {0}")]
+	Base64Error(#[from] base64::DecodeError),
+	#[error("Could not deserialize JSON: {0}")]
+	JSONError(#[from] serde_json::Error),
+}
+
+impl SyncToken {
+	pub fn new(sync_id: String, completed: bool, next_offset: usize) -> Self {
+		Self {
+			version: 1,
+			sync_id,
+			completed,
+			next_offset,
+		}
+	}
+
+	fn try_from_base64_str(s: &str) -> Result<Self, SyncTokenDeserializeError> {
+		let json_bytes = BASE64.decode(s)?;
+		let json = std::str::from_utf8(&json_bytes)?;
+		serde_json::from_str(json).map_err(Into::into)
+	}
+
+	pub fn try_from_header_value(
+		hv: &HeaderValue,
+	) -> Result<Self, SyncTokenDeserializeError> {
+		let s = hv.to_str()?;
+		Self::try_from_base64_str(s)
+	}
+
+	fn try_to_base64_string(self) -> Result<String, SyncTokenSerializeError> {
+		let json = serde_json::to_string(&self)?;
+		Ok(BASE64.encode(json))
+	}
+
+	pub fn try_to_header_value(self) -> Result<HeaderValue, SyncTokenSerializeError> {
+		let s = self.try_to_base64_string()?;
+		HeaderValue::from_str(s.as_ref()).map_err(Into::into)
+	}
+}
